refactor(memory): remove old code
refactor(memory): move page_info to separate header
