+++ kern/memory/pmm/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.16 FATAL_ERROR)
 
 target_sources(kernel
-        PRIVATE buddy_pmm.cc
+        #PRIVATE buddy_pmm.cc
         PRIVATE kmem.cc
         PRIVATE pmm.cc
         PRIVATE buddy_provider.cc
